♻️ refactor: move local components to app dir
