+++ app/document/[document_id]/components/Tab/index.tsx
